Persist Clusters and RecHits in flat ntuple
There's quite a bit more than the title says.

The Clusters created by CLUE are not first transformed in a SoA and
later saved in a flat ROOT tree, together with the original rechits used
to create the clusters. Local inspection of a file shows the everything
is correctly in place.

The CLUE clustering has been modified to properly take into account the
different configuration of the TestBeam setup. In particular the
parameters have been tuned differently for the EE and FH section and
for the later AHCAL volume. A quick inspection shows that, on average,
  CLUE can cluster ~93% of all the available energy of the rechits.